Move remove-props into util
